+++ packages/moonstone/VideoPlayer/MediaControls.js
@@ -7,6 +7,8 @@ import {SpotlightContainerDecorator, spotlightDefaultClass} from '@enact/spotlig
 import $L from '../internal/$L';
 import IconButton from '../IconButton';
 
+import {countReactChildren} from './util';
+
 import css from './VideoPlayer.less';
 
 const Container = SpotlightContainerDecorator({enterTo: ''}, 'div');
@@ -105,6 +107,19 @@ const MediaControls = kind({
 		 * @public
 		 */
 		moreButtonCloseLabel: PropTypes.string,
+
+		/**
+		 * The color of the underline beneath more icon button.
+		 *
+		 * This property accepts one of the following color names, which correspond with the
+		 * colored buttons on a standard remote control: `'red'`, `'green'`, `'yellow'`, `'blue'`
+		 *
+		 * @type {String}
+		 * @see {@link moonstone/IconButton.IconButtonBase.color}
+		 * @default 'blue'
+		 * @public
+		 */
+		moreButtonColor: PropTypes.oneOf([null, 'red', 'green', 'yellow', 'blue']),
 
 		/**
 		 * Sets the `disabled` state on the media "more" button.
@@ -296,6 +311,7 @@ const MediaControls = kind({
 		forwardIcon: 'forward',
 		jumpBackwardIcon: 'skipbackward',
 		jumpForwardIcon: 'skipforward',
+		moreButtonColor: 'blue',
 		pauseIcon: 'pause',
 		playIcon: 'play',
 		visible: true
@@ -334,8 +350,9 @@ const MediaControls = kind({
 			jumpButtonsDisabled,
 			jumpForwardIcon,
 			leftComponents,
-			moreButtonDisabled,
 			mediaDisabled,
+			moreButtonColor,
+			moreButtonDisabled,
 			moreDisabled,
 			moreIcon,
 			moreIconLabel,
@@ -366,14 +383,14 @@ const MediaControls = kind({
 		delete rest.visible;
 
 		return (
-			<div {...rest}>
+			<div {...rest} data-media-controls>
 				<div className={css.leftComponents}>{leftComponents}</div>
 				<div className={css.centerComponentsContainer}>
 					<div className={centerClassName}>
 						<Container className={css.mediaControls} spotlightDisabled={!moreDisabled || spotlightDisabled}>
 							{noJumpButtons ? null : <MediaButton aria-label={$L('Previous')} backgroundOpacity="translucent" disabled={mediaDisabled || jumpButtonsDisabled} onClick={onJumpBackwardButtonClick} spotlightDisabled={spotlightDisabled}>{jumpBackwardIcon}</MediaButton>}
 							{noRateButtons ? null : <MediaButton aria-label={$L('Rewind')} backgroundOpacity="translucent" disabled={mediaDisabled || rateButtonsDisabled} onClick={onBackwardButtonClick} spotlightDisabled={spotlightDisabled}>{backwardIcon}</MediaButton>}
-							<MediaButton aria-label={playPauseLabel} className={spotlightDefaultClass} backgroundOpacity="translucent" onClick={onPlayButtonClick} spotlightDisabled={spotlightDisabled}>{playPauseIcon}</MediaButton>
+							<MediaButton aria-label={playPauseLabel} className={spotlightDefaultClass} backgroundOpacity="translucent" disabled={mediaDisabled} onClick={onPlayButtonClick} spotlightDisabled={spotlightDisabled}>{playPauseIcon}</MediaButton>
 							{noRateButtons ? null : <MediaButton aria-label={$L('Fast Forward')} backgroundOpacity="translucent" disabled={mediaDisabled || rateButtonsDisabled} onClick={onForwardButtonClick} spotlightDisabled={spotlightDisabled}>{forwardIcon}</MediaButton>}
 							{noJumpButtons ? null : <MediaButton aria-label={$L('Next')} backgroundOpacity="translucent" disabled={mediaDisabled || jumpButtonsDisabled} onClick={onJumpForwardButtonClick} spotlightDisabled={spotlightDisabled}>{jumpForwardIcon}</MediaButton>}
 						</Container>
@@ -384,11 +401,12 @@ const MediaControls = kind({
 				</div>
 				<div className={css.rightComponents}>
 					{rightComponents}
-					{React.Children.count(children) ? (
+					{countReactChildren(children) ? (
 						<MediaButton
 							aria-label={moreIconLabel}
 							backgroundOpacity="translucent"
 							className={css.moreButton}
+							color={moreButtonColor}
 							disabled={moreButtonDisabled}
 							onTap={onToggleMore}
 							tooltipProps={{role: 'dialog'}}