+++ packages/moonstone/VirtualList/VirtualList.js
@@ -73,7 +73,12 @@ const VirtualGridList = kind({
 		 *
 		 * Usage:
 		 * ```
-		 * <VirtualGridList itemSize={{minWidth: ri.scale(180), minHeight: ri.scale(270)}} />
+		 * <VirtualGridList
+		 * 	itemSize={{
+		 * 		minWidth: ri.scale(180),
+		 * 		minHeight: ri.scale(270)
+		 * 	}}
+		 * />
 		 * ```
 		 *
 		 * @type {ui/VirtualList.gridListItemSizeShape}
@@ -151,7 +156,12 @@ const VirtualGridListNative = kind({
 		 *
 		 * Usage:
 		 * ```
-		 * <VirtualGridListNative itemSize={{minWidth: ri.scale(180), minHeight: ri.scale(270)}} />
+		 * <VirtualGridListNative
+		 * 	itemSize={{
+		 * 		minWidth: ri.scale(180),
+		 * 		minHeight: ri.scale(270)
+		 * 	}}
+		 * />
 		 * ```
 		 *
 		 * @type {ui/VirtualList.gridListItemSizeShape}