+++ packages/moonstone/VirtualList/VirtualListBase.js
@@ -54,25 +54,24 @@ const
 	JS = 'JS',
 	Native = 'Native';
 
+/**
+ * The base version of [VirtualListBase]{@link moonstone/VirtualList.VirtualListBase} and
+ * [VirtualListBaseNative]{@link moonstone/VirtualList.VirtualListBaseNative}.
+ *
+ * @class VirtualListCore
+ * @memberof moonstone/VirtualList
+ * @ui
+ * @public
+ */
 const VirtualListBaseFactory = (type) => {
 	const UiBase = (type === JS) ? UiVirtualListBase : UiVirtualListBaseNative;
 
 	return class VirtualListCore extends Component {
-		static displayName = 'VirtualListBase'
+		/* No displayName here. We set displayName to returned components of this factory function. */
 
-		static propTypes = /** @lends moonstone/VirtualList.VirtualList.prototype */ {
+		static propTypes = /** @lends moonstone/VirtualList.VirtualListCore.prototype */ {
 			/**
-			 * The `render` function for an item of the list receives the following parameters:
-			 * - `data` is for accessing the supplied `data` property of the list.
-			 * > NOTE: In most cases, it is recommended to use data from redux store instead of using
-			 * is parameters due to performance optimizations.
-			 *
-			 * @param {Object} event
-			 * @param {Number} event.data-index It is required for Spotlight 5-way navigation. Pass to the root element in the component.
-			 * @param {Number} event.index The index number of the componet to render
-			 * @param {Number} event.key It MUST be passed as a prop to the root element in the component for DOM recycling.
-			 *
-			 * Data manipulation can be done in this function.
+			 * The `render` function called for each item in the list.
 			 *
 			 * > NOTE: The list does NOT always render a component whenever its render function is called
 			 * due to performance optimization.
@@ -80,15 +79,21 @@ const VirtualListBaseFactory = (type) => {
 			 * Usage:
 			 * ```
 			 * renderItem = ({index, ...rest}) => {
-			 *		delete rest.data;
+			 * 	delete rest.data;
 			 *
-			 *		return (
-			 *			<MyComponent index={index} {...rest} />
-			 *		);
+			 * 	return (
+			 * 		<MyComponent index={index} {...rest} />
+			 * 	);
 			 * }
 			 * ```
 			 *
 			 * @type {Function}
+			 * @param {Object} event
+			 * @param {Number} event.data-index It is required for Spotlight 5-way navigation. Pass to the root element in the component.
+			 * @param {Number} event.index The index number of the component to render
+			 * @param {Number} event.key It MUST be passed as a prop to the root element in the component for DOM recycling.
+			 *
+			 * @required
 			 * @public
 			 */
 			itemRenderer: PropTypes.func.isRequired,
@@ -97,6 +102,7 @@ const VirtualListBaseFactory = (type) => {
 			 * The render function for the items.
 			 *
 			 * @type {Function}
+			 * @required
 			 * @private
 			 */
 			itemsRenderer: PropTypes.func.isRequired,
@@ -109,6 +115,32 @@ const VirtualListBaseFactory = (type) => {
 			 * @private
 			 */
 			initUiChildRef: PropTypes.func,
+
+			/**
+			 * The Function that returns `true` if the item at the index is disabled.
+			 * It is used to navigate a list properly with 5 way keys, page up key,
+			 * and page down key. If it is not supplied, it assumes that no items are disabled.
+			 *
+			 * Usage:
+			 * ```
+			 * isItemDisabled = (index) => (this.items[index].disabled)
+			 * render = () => {
+			 * 	return (
+			 * 		<VirtualList
+			 * 			dataSize={this.items.length}
+			 * 			isItemDisabled={isItemDisabled}
+			 * 			itemRenderer={this.renderItem}
+			 * 			itemSize={this.itemSize}
+			 * 		/>
+			 * 	);
+			 * }
+			 * ```
+			 *
+			 * @type {Function}
+			 * @param {Number} index
+			 * @public
+			 */
+			isItemDisabled: PropTypes.func,
 
 			/**
 			 * `true` if rtl, `false` if ltr.
@@ -138,9 +170,9 @@ const VirtualListBaseFactory = (type) => {
 		}
 
 		componentDidMount () {
-			if (type === JS) {
-				const containerNode = this.uiRef.containerRef;
+			const containerNode = this.uiRef.containerRef;
 
+			if (type === JS) {
 				// prevent native scrolling by Spotlight
 				this.preventScroll = () => {
 					containerNode.scrollTop = 0;
@@ -149,14 +181,11 @@ const VirtualListBaseFactory = (type) => {
 
 				if (containerNode && containerNode.addEventListener) {
 					containerNode.addEventListener('scroll', this.preventScroll);
-					containerNode.addEventListener('keydown', this.onKeyDown);
 				}
-			} else {
-				const contentNode = this.uiRef.contentRef;
+			}
 
-				if (contentNode && contentNode.addEventListener) {
-					contentNode.addEventListener('keydown', this.onKeyDown);
-				}
+			if (containerNode && containerNode.addEventListener) {
+				containerNode.addEventListener('keydown', this.onKeyDown);
 			}
 		}
 
@@ -171,20 +200,18 @@ const VirtualListBaseFactory = (type) => {
 		}
 
 		componentWillUnmount () {
-			if (type === JS) {
-				const containerNode = this.uiRef.containerRef;
+			const containerNode = this.uiRef.containerRef;
 
+			if (type === JS) {
 				// remove a function for preventing native scrolling by Spotlight
 				if (containerNode && containerNode.removeEventListener) {
 					containerNode.removeEventListener('scroll', this.preventScroll);
 					containerNode.removeEventListener('keydown', this.onKeyDown);
 				}
-			} else {
-				const contentNode = this.uiRef.contentRef;
+			}
 
-				if (contentNode && contentNode.removeEventListener) {
-					contentNode.removeEventListener('keydown', this.onKeyDown);
-				}
+			if (containerNode && containerNode.removeEventListener) {
+				containerNode.removeEventListener('keydown', this.onKeyDown);
 			}
 
 			this.setContainerDisabled(false);
@@ -198,11 +225,10 @@ const VirtualListBaseFactory = (type) => {
 		restoreLastFocused = false
 
 		setContainerDisabled = (bool) => {
-			const {spotlightId} = this.props;
-			const containerNode = document.querySelector(`[data-spotlight-id="${spotlightId}"]`);
+			const contentNode = this.uiRef.contentRef;
 
-			if (containerNode) {
-				containerNode.setAttribute(dataContainerDisabledAttribute, bool);
+			if (contentNode) {
+				contentNode.setAttribute(dataContainerDisabledAttribute, bool);
 
 				if (bool) {
 					document.addEventListener('keydown', this.handleGlobalKeyDown, {capture: true});
@@ -218,7 +244,8 @@ const VirtualListBaseFactory = (type) => {
 
 		findSpottableItem = (indexFrom, indexTo) => {
 			const
-				{data, dataSize} = this.uiRef.props,
+				{isItemDisabled} = this.props,
+				{dataSize} = this.uiRef.props,
 				safeIndexFrom = clamp(0, dataSize - 1, indexFrom),
 				safeIndexTo = clamp(-1, dataSize, indexTo),
 				delta = (indexFrom < indexTo) ? 1 : -1;
@@ -229,7 +256,7 @@ const VirtualListBaseFactory = (type) => {
 
 			if (safeIndexFrom !== safeIndexTo) {
 				for (let i = safeIndexFrom; i !== safeIndexTo; i += delta) {
-					if (data[i] && data[i].disabled === false) {
+					if (isItemDisabled(i) === false) {
 						return i;
 					}
 				}
@@ -240,7 +267,8 @@ const VirtualListBaseFactory = (type) => {
 
 		getIndexToScrollDisabled = (direction, currentIndex) => {
 			const
-				{data, dataSize, spacing} = this.uiRef.props,
+				{isItemDisabled} = this.props,
+				{dataSize, spacing} = this.uiRef.props,
 				{dimensionToExtent, primary} = this.uiRef,
 				{findSpottableItem} = this,
 				{firstVisibleIndex, lastVisibleIndex} = this.uiRef.moreInfo,
@@ -292,7 +320,7 @@ const VirtualListBaseFactory = (type) => {
 					distance,
 					index;
 				for (let i = firstIndexInExtent; i <= lastIndexInExtent; ++i) {
-					if (data[i] && !data[i].disabled) {
+					if (!isItemDisabled(i)) {
 						distance = Math.abs(currentPosInExtent - i % dimensionToExtent);
 						if (distance < minDistance) {
 							minDistance = distance;
@@ -329,12 +357,12 @@ const VirtualListBaseFactory = (type) => {
 
 		scrollToNextItem = ({direction, focusedItem}) => {
 			const
-				{data} = this.uiRef.props,
-				focusedIndex = Number.parseInt(focusedItem.getAttribute(dataIndexAttribute)),
-				{firstVisibleIndex, lastVisibleIndex} = this.uiRef.moreInfo;
+				{isItemDisabled} = this.props,
+				{firstIndex, numOfItems} = this.uiRef.state,
+				focusedIndex = Number.parseInt(focusedItem.getAttribute(dataIndexAttribute));
 			let indexToScroll = -1;
 
-			if (Array.isArray(data) && data.some((item) => item.disabled)) {
+			if (isItemDisabled) {
 				indexToScroll = this.getIndexToScrollDisabled(direction, focusedIndex);
 			} else {
 				indexToScroll = this.getIndexToScroll(direction, focusedIndex);
@@ -345,13 +373,7 @@ const VirtualListBaseFactory = (type) => {
 					isRtl = this.props.rtl,
 					isForward = (direction === 'down' || isRtl && direction === 'left' || !isRtl && direction === 'right');
 
-				if (type === JS) {
-					// To prevent item positioning issue, make all items to be rendered.
-					this.uiRef.updateFrom = null;
-					this.uiRef.updateTo = null;
-				}
-
-				if (firstVisibleIndex <= indexToScroll && indexToScroll <= lastVisibleIndex) {
+				if (firstIndex <= indexToScroll && indexToScroll < firstIndex + numOfItems) {
 					const node = this.uiRef.containerRef.querySelector(`[data-index='${indexToScroll}'].spottable`);
 
 					if (node) {
@@ -363,8 +385,8 @@ const VirtualListBaseFactory = (type) => {
 						Spotlight.pause();
 					}
 					focusedItem.blur();
+					this.nodeIndexToBeFocused = this.lastFocusedIndex = indexToScroll;
 				}
-				this.nodeIndexToBeFocused = this.lastFocusedIndex = indexToScroll;
 				this.uiRef.props.cbScrollTo({index: indexToScroll, stickTo: isForward ? 'end' : 'start', animate: false});
 			}
 
@@ -401,13 +423,17 @@ const VirtualListBaseFactory = (type) => {
 
 		jumpToSpottableItem = (keyCode, target) => {
 			const
-				{cbScrollTo, data, dataSize} = this.uiRef.props,
+				{isItemDisabled} = this.props,
+				{cbScrollTo, dataSize} = this.uiRef.props,
 				{firstIndex, numOfItems} = this.uiRef.state,
 				{isPrimaryDirectionVertical} = this.uiRef,
 				rtl = this.props.rtl,
 				currentIndex = Number.parseInt(target.getAttribute(dataIndexAttribute));
 
-			if (!data || !Array.isArray(data) || !data[currentIndex] || data[currentIndex].disabled) {
+			if (
+				!isItemDisabled || // It is the case that there is no disabled items in a list.
+				isItemDisabled(currentIndex) // If the currnet index item is disabled, it means that all items in a list are disabled.
+			) {
 				return false;
 			}
 
@@ -427,12 +453,12 @@ const VirtualListBaseFactory = (type) => {
 
 			if (isForward) {
 				// See if the next item is spottable then delegate scroll to onFocus handler
-				if (currentIndex < dataSize - 1 && !data[currentIndex + 1].disabled) {
+				if (currentIndex < dataSize - 1 && !isItemDisabled(currentIndex + 1)) {
 					return false;
 				}
 
 				for (let i = currentIndex + 2; i < dataSize; i++) {
-					if (!data[i].disabled) {
+					if (!isItemDisabled(i)) {
 						nextIndex = i;
 						break;
 					}
@@ -445,12 +471,12 @@ const VirtualListBaseFactory = (type) => {
 				}
 			} else if (isBackward) {
 				// See if the next item is spottable then delegate scroll to onFocus handler
-				if (currentIndex > 0 && !data[currentIndex - 1].disabled) {
+				if (currentIndex > 0 && !isItemDisabled(currentIndex - 1)) {
 					return false;
 				}
 
 				for (let i = currentIndex - 2; i >= 0; i--) {
-					if (!data[i].disabled) {
+					if (!isItemDisabled(i)) {
 						nextIndex = i;
 						break;
 					}
@@ -494,9 +520,7 @@ const VirtualListBaseFactory = (type) => {
 
 			this.isScrolledBy5way = false;
 			if (getDirection(keyCode)) {
-				if (type === Native) {
-					ev.preventDefault();
-				}
+				ev.preventDefault();
 				this.setSpotlightContainerRestrict(keyCode, target);
 				this.isScrolledBy5way = this.jumpToSpottableItem(keyCode, target);
 			}
@@ -526,9 +550,6 @@ const VirtualListBaseFactory = (type) => {
 
 			if (Spotlight.isPaused()) {
 				Spotlight.resume();
-				if (type === JS) {
-					this.forceUpdate();
-				}
 			}
 			this.focusOnNode(item);
 			this.nodeIndexToBeFocused = null;
@@ -539,7 +560,7 @@ const VirtualListBaseFactory = (type) => {
 			if (ref) {
 				if (type === JS) {
 					this.focusOnItem(index);
-				} else if (type === Native) {
+				} else {
 					// If focusing the item of VirtuallistNative, `onFocus` in Scrollable will be called.
 					// Then VirtualListNative tries to scroll again differently from VirtualList.
 					// So we would like to skip `focus` handling when focusing the item as a workaround.
@@ -640,9 +661,7 @@ const VirtualListBaseFactory = (type) => {
 						node.blur();
 					}
 				}
-				if (type === JS) {
-					this.nodeIndexToBeFocused = null;
-				}
+				this.nodeIndexToBeFocused = null;
 				this.lastFocusedIndex = focusedIndex;
 
 				if (primary.clientSize >= primary.itemSize) {
@@ -713,6 +732,7 @@ const VirtualListBaseFactory = (type) => {
 
 			delete rest.spotlightId;
 			delete rest.initUiChildRef;
+			delete rest.isItemDisabled;
 
 			return (
 				<UiBase
@@ -748,9 +768,10 @@ const VirtualListBaseFactory = (type) => {
  * @memberof moonstone/VirtualList
  * @extends ui/VirtualList.VirtualListBase
  * @ui
- * @private
+ * @public
  */
 const VirtualListBase = VirtualListBaseFactory(JS);
+VirtualListBase.displayName = 'VirtualListBase';
 
 /**
  * A Moonstone-styled base component for [VirtualListNative]{@link moonstone/VirtualList.VirtualListNative} and
@@ -763,6 +784,7 @@ const VirtualListBase = VirtualListBaseFactory(JS);
  * @private
  */
 const VirtualListBaseNative = VirtualListBaseFactory(Native);
+VirtualListBaseNative.displayName = 'VirtualListBaseNative';
 
 const ScrollableVirtualList = ({role, ...rest}) => ( // eslint-disable-line react/jsx-no-bind
 	<Scrollable
@@ -790,12 +812,12 @@ const ScrollableVirtualList = ({role, ...rest}) => ( // eslint-disable-line reac
 	/>
 );
 
-ScrollableVirtualList.propTypes = /** @lends moonstone/VirtualList.VirtualList.prototype */ {
+ScrollableVirtualList.propTypes = /** @lends moonstone/VirtualList.VirtualListBase.prototype */ {
 	/**
 	 * Aria role.
 	 *
 	 * @type {String}
-	 * @private
+	 * @public
 	 */
 	role: PropTypes.string
 };
@@ -826,7 +848,7 @@ const ScrollableVirtualListNative = ({role, ...rest}) => (
 	/>
 );
 
-ScrollableVirtualListNative.propTypes = /** @lends moonstone/VirtualList.VirtualListNative.prototype */ {
+ScrollableVirtualListNative.propTypes = /** @lends moonstone/VirtualList.VirtualListBaseNative.prototype */ {
 	/**
 	 * Aria role.
 	 *