+++ packages/ui/Scrollable/Scrollable.js
@@ -488,7 +488,7 @@ class ScrollableBase extends Component {
 	onDragStart = (ev) => {
 		this.stop();
 		this.isDragging = true;
-		this.dragStartX = this.scrollLeft + ev.x;
+		this.dragStartX = this.scrollLeft + this.getRtlX(ev.x);
 		this.dragStartY = this.scrollTop + ev.y;
 	}